+++ CHANGELOG.rst
@@ -22,6 +22,7 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 * Created the project using QUBE
 * Implement queries for affiliation creation in database
 * Implement use case _List Packages_
+* Dynamic cost overview upon offer creation
 
 **Fixed**